src/app/pages/landing-page/landing-page.component.ts
import { Component } from '@angular/core';
import { Router } from '@angular/router';
import {MatButtonModule} from '@angular/material/button';

@Component({
  selector: 'app-landing-page',
  standalone: true,
  imports: [MatButtonModule],
  templateUrl: './landing-page.component.html',
  styleUrl: './landing-page.component.scss'
})
export class LandingPageComponent {
  constructor (private router:Router) {}

  toLogin(){
    this.router.navigate(['login']);
  }

  toRegister(){
    this.router.navigate(['register']);
  }
}
